Fabric framework & new social tab
Fabric can be used for its REST API and other things. I made another
tab that is an “Intro” page to the social side of the app. Twitter
Authentication is the only thing implemented (does nothing after
logging in).(Adding new UIViewControllers and buttons to connect them
to each other as a way of getting around).