Took out login in twitter for now. replaced swipe right feature for back button.
we can implement the login for the social side back on later if we want
to expand on the extra features logging in users will give us. The
swipe right feature needed work so I decided to take it out and to make
it simpler (via using a back button instead) to push and pop views on
the stack for now.