Fixed Compile errors
It runs again. This is without the settings and Social features
attached. We can’t just make calls with the ‘client_id’ for Instagram
so we need to go through authorization. Settings is just a template.
UI/UX needs to be worked on. I also have to clean up the code and repo
some more because of the recent updates.